renamed kuzzleSubscribeResult.js -> SubscribeResult.js
